refactor: cases
