+++ queue/src/fiskaltrust.Middleware.Localization.v2/Models/Cases/ES/ChargeItemCaseNatureOfVatES.cs
@@ -1,6 +1,6 @@
 using fiskaltrust.ifPOS.v2.Cases;
 
-namespace fiskaltrust.Middleware.Localization.v2.Models.Cases;
+namespace fiskaltrust.Middleware.Localization.v2.Models.Cases.ES;
 
 public enum ChargeItemCaseNatureOfVatES : long
 {
