+++ queue/src/fiskaltrust.Middleware.Localization.v2/Models/Cases/PT/ChargeItemCaseNatureOfVatPT.cs
@@ -1,6 +1,6 @@
 using fiskaltrust.ifPOS.v2.Cases;
 
-namespace fiskaltrust.Middleware.Localization.v2.Models.Cases;
+namespace fiskaltrust.Middleware.Localization.v2.Models.Cases.PT;
 
 public enum ChargeItemCaseNatureOfVatPT
 {
