+++ queue/src/fiskaltrust.Middleware.Localization.v2/Models/Cases/PT/TaxExemptionPT.cs
@@ -1,4 +1,4 @@
-namespace fiskaltrust.Middleware.Localization.v2.Models.Cases;
+namespace fiskaltrust.Middleware.Localization.v2.Models.Cases.PT;
 
 public enum TaxExemptionCodePT
 {
