+++ queue/src/fiskaltrust.Middleware.Localization.v2/Validation/Rules/ES/ChargeItemValidations.cs
@@ -1,6 +1,6 @@
 using fiskaltrust.ifPOS.v2;
 using fiskaltrust.ifPOS.v2.Cases;
-using fiskaltrust.Middleware.Localization.v2.Models.Cases;
+using fiskaltrust.Middleware.Localization.v2.Models.Cases.ES;
 using FluentValidation;
 
 namespace fiskaltrust.Middleware.Localization.v2.Validation.Rules.ES;
